+++ lib/components/exports.ts
@@ -5,4 +5,3 @@ export * from "./Layout/exports"
 export * from "./Navigation/exports"
 export * from "./Overlays/exports"
 export * from "./Utilities/exports"
-export * from "./Widgets/exports"
