+++ lib/experimental/Widgets/CategoryBarSection/index.tsx
