+++ lib/experimental/Widgets/Charts/BarChartWidget/index.stories.tsx
