+++ lib/experimental/Widgets/Charts/BarChartWidget/index.tsx
