+++ lib/experimental/Widgets/Charts/VerticalBarChartWidget/index.stories.tsx
