+++ lib/experimental/Widgets/Charts/VerticalBarChartWidget/index.tsx
