+++ src/playground/Pages/Tabs/index.stories.tsx
@@ -1,14 +1,14 @@
 import { AreaChartProps } from "@/components/Charts/AreaChart"
-import { AreaChartWidget } from "@/components/Widgets/Charts/AreaChartWidget"
-import AreaChartWidgetStoriesMeta from "@/components/Widgets/Charts/AreaChartWidget/index.stories"
-import { ComposeChartContainerProps } from "@/components/Widgets/Charts/ChartContainer"
-import { LineChartWidget } from "@/components/Widgets/Charts/LineChartWidget"
-import LineChartWidgetStoriesMeta from "@/components/Widgets/Charts/LineChartWidget/index.stories"
-import { PieChartWidget } from "@/components/Widgets/Charts/PieChartWidget"
-import PieChartWidgetStoriesMeta from "@/components/Widgets/Charts/PieChartWidget/index.stories"
-import { VerticalBarChartWidget } from "@/components/Widgets/Charts/VerticalBarChartWidget"
-import VerticalBarChartWidgetStoriesMeta from "@/components/Widgets/Charts/VerticalBarChartWidget/index.stories"
-import { Dashboard } from "@/components/Widgets/Dashboard"
+import { AreaChartWidget } from "@/experimental/Widgets/Charts/AreaChartWidget"
+import AreaChartWidgetStoriesMeta from "@/experimental/Widgets/Charts/AreaChartWidget/index.stories"
+import { ComposeChartContainerProps } from "@/experimental/Widgets/Charts/ChartContainer"
+import { LineChartWidget } from "@/experimental/Widgets/Charts/LineChartWidget"
+import LineChartWidgetStoriesMeta from "@/experimental/Widgets/Charts/LineChartWidget/index.stories"
+import { PieChartWidget } from "@/experimental/Widgets/Charts/PieChartWidget"
+import PieChartWidgetStoriesMeta from "@/experimental/Widgets/Charts/PieChartWidget/index.stories"
+import { VerticalBarChartWidget } from "@/experimental/Widgets/Charts/VerticalBarChartWidget"
+import VerticalBarChartWidgetStoriesMeta from "@/experimental/Widgets/Charts/VerticalBarChartWidget/index.stories"
+import { Dashboard } from "@/experimental/Widgets/Dashboard"
 import { User } from "@/icons"
 import type { Meta, StoryObj } from "@storybook/react"
 import { Tabs } from "."
